Added python command to get list of spawners sorted by proximity to a given point
Added python command to fire input on a specific entity
Taking fall damage after landing on a player now transfers some damage to them
Moved python newer python entity commands from GEUtil to GEEntity

+++ game/sdk/shared/takedamageinfo.cpp
@@ -55,7 +55,7 @@ void CTakeDamageInfo::Init( CBaseEntity *pInflictor, CBaseEntity *pAttacker, CBa
 	m_iDamageCustom = iCustomDamage;
 
 #ifdef GE_DLL
-    m_iDamageCap = INT_MAX; // Can do unlimited damage by default.
+    m_iDamageCap = SHRT_MAX; // Can do unlimited damage by default.
 
 	m_iDamageStats = 0;
 #endif
